+++ glyph/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+* Add `OutlinedGlyph::draw_using` allowing reusing of `Rasterizer` allocation.
 * Update _ttf-parser_ to `0.10`.
 
 # 0.2.7